Updated to include use of new batch files
As part of the newly re-worked stopsigparser, we have also created a helpful utility called 'Make Folders.bat' that makes the folders required for the parser right in the user's directory of choice. This enhances the deployment, given that the user is no longer required to make the exact file structure required for operation, a big upgrade over our previous deployable version of stopsigparser. It also includes two new .bat scripts that allow for either 1) collection of data alone or 2) collection of both graphs and data.